Extract printInterface and printCallExpression
